Fix audio playback from cue/bin file

+++ src/interfaces/3DO.c
@@ -18,7 +18,7 @@
 
 #include "write.pio.h"
 
-extern bool readBlock(uint32_t start, uint16_t nb_block, uint8_t block_size, uint8_t *buffer);
+extern bool readBlock(uint32_t start, uint16_t nb_block, uint16_t block_size, uint8_t *buffer);
 extern bool readSubQChannel(uint8_t *buffer);
 extern bool driveEject(bool eject);
 extern bool block_is_ready();
